subscriber/requirements.txt
Django==3.2.15
psycopg2_binary==2.9.3
python-dotenv==0.15.0
paho-mqtt==1.6.1
flake8